requirements.txt
astroid==2.2.5
autopep8==1.4.3
awsebcli==3.14.13
blessed==1.15.0
botocore==1.12.123
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-rest-framework==0.1.0
djangorestframework==3.9.2
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.3
flake8==3.7.7
future==0.16.0
idna==2.7
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.1.6
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
pathspec==0.5.9
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.1
